practice/chapter13/code/requirements.txt
# Chapter 13: RAG and Prompt Engineering
torch>=2.0.0
transformers>=4.36.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
langchain>=0.1.0
langchain-community>=0.0.10
langchain-huggingface>=0.0.1
chromadb>=0.4.0
numpy>=1.24.0
tqdm>=4.65.0
